yoseop_1/requirements.txt
# AI 면접 시스템 의존성 패키지

# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/LLM
openai==1.3.7
tiktoken==0.5.1

# 데이터베이스
supabase==2.17.0
psycopg2-binary==2.9.9

# 데이터 처리
pandas==2.1.3
numpy==1.25.2
pydantic==2.11.7

# 문서 처리
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# HTTP 요청
httpx==0.27.0
requests==2.31.0

# 환경 변수
python-dotenv==1.0.0

# 로깅
loguru==0.7.2

# 유틸리티
python-dateutil==2.9.0
uuid==1.30

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# CORS 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 기타
typing-extensions==4.14.1
pathlib==1.0.1